+++ main.go
@@ -5,10 +5,8 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"time"
 
 	"github.com/fatih/color"
-	gosentry "github.com/getsentry/sentry-go"
 	"github.com/joho/godotenv"
 	"github.com/sirupsen/logrus"
 	"github.com/snyk/driftctl/build"
@@ -18,7 +16,6 @@ import (
 	"github.com/snyk/driftctl/pkg/cmd/scan"
 	"github.com/snyk/driftctl/pkg/config"
 	"github.com/snyk/driftctl/pkg/version"
-	"github.com/snyk/driftctl/sentry"
 )
 
 func init() {
@@ -59,27 +56,10 @@ func run() int {
 		}()
 	}
 
-	// Handle panic and log them to sentry if error reporting is enabled
-	defer func() {
-		if cmd.IsReportingEnabled(&driftctlCmd.Command) {
-			err := recover()
-			if err != nil {
-				gosentry.CurrentHub().Recover(err)
-				flushSentry()
-				logrus.Fatalf("Captured panic: %s", err)
-				os.Exit(scan.ExitError)
-			}
-			flushSentry()
-		}
-	}()
-
 	if _, err := driftctlCmd.ExecuteC(); err != nil {
 		var notInSync cmderrors.InfrastructureNotInSync
 		if errors.As(err, &notInSync) {
 			return scan.ExitNotInSync
-		}
-		if cmd.IsReportingEnabled(&driftctlCmd.Command) {
-			sentry.CaptureException(err)
 		}
 		_, _ = fmt.Fprintln(os.Stderr, color.RedString("%s", err))
 		return scan.ExitError
@@ -94,10 +74,4 @@ func run() int {
 	}
 
 	return scan.ExitInSync
-}
-
-func flushSentry() {
-	ttl := 60 * time.Second
-	ok := gosentry.Flush(ttl)
-	logrus.WithField("timeout", ttl).WithField("success", ok).Debug("Flushed Sentry events")
 }