+++ pkg/cmd/driftctl.go
@@ -7,7 +7,6 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/snyk/driftctl/build"
 	"github.com/snyk/driftctl/pkg"
-	"github.com/snyk/driftctl/sentry"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
@@ -43,11 +42,7 @@ func NewDriftctlCmd(build build.Interface) *DriftctlCmd {
 			Use:   "driftctl <command> [flags]",
 			Short: "Driftctl CLI",
 			PersistentPreRunE: func(cmd *cobra.Command, _ []string) error {
-				err := bindEnvToFlags(cmd)
-				if err != nil {
-					return err
-				}
-				return handleReporting(cmd)
+				return bindEnvToFlags(cmd)
 			},
 			Long:          "Detect, track and alert on infrastructure drift.",
 			SilenceErrors: true,
@@ -66,9 +61,7 @@ func NewDriftctlCmd(build build.Interface) *DriftctlCmd {
 	cmd.PersistentFlags().BoolP("help", "h", false, "Display help for command")
 	if cmd.build.IsUsageReportingEnabled() {
 		cmd.PersistentFlags().BoolP("no-version-check", "", false, "Disable the version check")
-		cmd.PersistentFlags().BoolP("disable-telemetry", "", false, "Disable telemetry")
 	}
-	cmd.PersistentFlags().BoolP("send-crash-report", "", false, "Enable error reporting. Crash data will be sent to us via Sentry.\nWARNING: may leak sensitive data (please read the documentation for more details)\nThis flag should be used only if an error occurs during execution")
 
 	cmd.AddCommand(NewScanCmd(&pkg.ScanOptions{}))
 	cmd.AddCommand(NewFmtCmd(&pkg.FmtOptions{}))
@@ -94,22 +87,6 @@ func (driftctlCmd DriftctlCmd) ShouldCheckVersion() bool {
 	hasCompletionCmd := contains(os.Args[1:], "completion")
 	isHelp := contains(os.Args[1:], "help") || contains(os.Args[1:], "--help") || contains(os.Args[1:], "-h")
 	return driftctlCmd.build.IsRelease() && driftctlCmd.build.IsUsageReportingEnabled() && !hasVersionCmd && !hasCompletionCmd && !noVersionCheckVal && !isHelp && !noVersionCheckEnv
-}
-
-// IsReportingEnabled returns whether error reporting is enabled.
-func IsReportingEnabled(cmd *cobra.Command) bool {
-	enableReporting, err := cmd.Flags().GetBool("send-crash-report")
-	if err != nil {
-		return false
-	}
-	return enableReporting
-}
-
-func handleReporting(cmd *cobra.Command) error {
-	if IsReportingEnabled(cmd) {
-		return sentry.Initialize()
-	}
-	return nil
 }
 
 // Iterate over command flags