+++ pkg/cmd/scan.go
@@ -16,7 +16,6 @@ import (
 	"github.com/mitchellh/go-homedir"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
-	"github.com/snyk/driftctl/build"
 	"github.com/snyk/driftctl/enumeration/alerter"
 	"github.com/snyk/driftctl/enumeration/remote"
 	"github.com/snyk/driftctl/enumeration/remote/aws"
@@ -32,7 +31,6 @@ import (
 	"github.com/snyk/driftctl/pkg/memstore"
 	dctlresource "github.com/snyk/driftctl/pkg/resource"
 	"github.com/snyk/driftctl/pkg/resource/schemas"
-	"github.com/snyk/driftctl/pkg/telemetry"
 	"github.com/snyk/driftctl/pkg/terraform/hcl"
 	"github.com/spf13/cobra"
 
@@ -116,7 +114,6 @@ func NewScanCmd(opts *pkg.ScanOptions) *cobra.Command {
 			}
 
 			opts.Quiet, _ = cmd.Flags().GetBool("quiet")
-			opts.DisableTelemetry, _ = cmd.Flags().GetBool("disable-telemetry")
 
 			opts.ConfigDir, _ = cmd.Flags().GetString("config-dir")
 
@@ -370,7 +367,6 @@ func scanRun(opts *pkg.ScanOptions) error {
 
 	analysis.ProviderVersion = opts.ProviderVersion
 	analysis.ProviderName = opts.To
-	store.Bucket(memstore.TelemetryBucket).Set("provider_name", analysis.ProviderName)
 
 	// Categorize unmanaged resources for AWS providers
 	if strings.HasPrefix(opts.To, "aws") {
@@ -451,11 +447,6 @@ func scanRun(opts *pkg.ScanOptions) error {
 
 	globaloutput.Printf(color.WhiteString("Scan duration: %s\n", analysis.Duration.Round(time.Second)))
 	globaloutput.Printf(color.WhiteString("Provider version used to scan: %s. Use --tf-provider-version to use another version.\n"), opts.ProviderVersion)
-
-	if !opts.DisableTelemetry {
-		tl := telemetry.NewTelemetry(&build.Build{})
-		tl.SendTelemetry(store.Bucket(memstore.TelemetryBucket))
-	}
 
 	if !analysis.IsSync() {
 		return cmderrors.InfrastructureNotInSync{}